refactor: toast Atom 분리
